reset back to m1/compressed commitment working and then keep all new network robustness cleanup
