--- Sources/Utils/MobileCoin-Bridging-Header.h
@@ -1,18 +1,0 @@
-//
-//  Copyright (c) 2020-2021 MobileCoin. All rights reserved.
-//
-
-#ifndef MobileCoin_Bridging_Header_h
-#define MobileCoin_Bridging_Header_h
-
-
-#endif /* MobileCoin_Bridging_Header_h */
-
-#ifndef TestBridgingHeader_h
-#define TestBridgingHeader_h
-
-
-
-#endif /* TestBridgingHeader_h */
-
-#include "TestBridgingHeader.h"
