--- Sources/Utils/TestBridgingHeader.h
@@ -1,11 +1,0 @@
-//
-//  Copyright (c) 2020-2021 MobileCoin. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-
-@interface TestBridgingHeader : NSObject
-
-- (id)init;
-
-@end
