--- Sources/Utils/TestBridgingHeader.m
@@ -1,16 +1,0 @@
-//
-//  Copyright (c) 2020-2021 MobileCoin. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-
-#include "TestBridgingHeader.h"
-
-@implementation TestBridgingHeader
-
-- (id)init
-{
-    return self;
-}
-
-@end
